Extract to src/components/managerlib/pathutils.ts
